Move order controller to customer folder
