+++ packages/Webkul/Shop/src/Routes/customer-routes.php
@@ -10,7 +10,7 @@ use Webkul\Shop\Http\Controllers\Customer\ResetPasswordController;
 use Webkul\Shop\Http\Controllers\Customer\SessionController;
 use Webkul\Shop\Http\Controllers\Customer\WishlistController;
 use Webkul\Shop\Http\Controllers\DownloadableProductController;
-use Webkul\Shop\Http\Controllers\OrderController;
+use Webkul\Shop\Http\Controllers\Customer\OrderController;
 use Webkul\Shop\Http\Controllers\ReviewController;
 
 Route::group(['middleware' => ['locale', 'theme', 'currency']], function () {
